v0.0.5 BUG FIX
Fixed a bug in the calculation of the impurity Green's functions.
The real part of the GF had a wrong sign compared to the Lanczos code,
this was related to a wrong sign in the excitation energy \Delta E in
the spectral sum. Found and Checked by F.Grandi in the 2bands Bethe
code.